requirements.txt
httpx
ipython
pydantic
langgraph
langchain-core
langchain-openai
langchain-tavily
python-dotenv
requests
pyppeteer
mypy